requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
flask
gunicorn
python-dotenv
os